🎉 add: parsing.h & move parsing functions to their own file
